[#181] Add Project Documentation
This commit adds the Official Project Documentation, using Sphinx and
reStructuredText.

The topics covered include a documentation and application overview,
setup and installation, a user guide, information on contributing to the
project, design specifications and the API documentation.

plantUML is used to generate UML diagrams for the documentation, such as
Activity and Usecase Diagrams. plantUML requires java to be installed.

Pencil is used to create the wireframes for the Design Specifications.

Closes #181: Write Program Specifications