[#116] Add Inactive Account Functionality
Accounts that are marked inactive will now not be shown in entry or
dropdown menus. These Accounts can still be viewed in the Chart of
Accounts.

This feature is the planned way to mark an Account for future deletion.
An Account can be marked inactive so no new Transactions will be created
for it. Creating a new Fiscl Year will prune all the old Transactions,
allowing the user to delete the Account.

* Add an `active` method to the AccountManager to retrieve a queryset
  containing only Accounts marked active.

* Modify the Quck Account and Bank Dropdowns to display only Accounts
  marked active.
* Modify applicable Entry/Transaction Forms to display only Accounts
  marked active in the Account Widgets.

* Fix a bug in the AccountReconcileViewTests caused by previous commit.

Fixes #116: Making an Account Inactive has no Effect