[#217] Reorganize Package Structure
The accounts package was overflowing with everything in the application.
These changes correct this by breaking parts of the accounts package
into different packages.

The new packages are accounts, core, entries, events and fiscalyears.
The core package holds items used in the majority of packages such as
template filters and forms.

* Add the core, entries, events and fiscalyears packages.

* Modify the root URL to use the show_accounts_chart view.

* Refactor the process_quick_ account, bank and event functions into a
  single process_quick_search function in the core.core module.
* Refactor model managers into a separate module.

Closes #136: Move Model Managers into managers.py

Closes #217: Split up the accounts package