[#284] Fix Reconcile Page's Reconciled Balance
Fix a bug in the reconcile_account view where an invalid formset would
cause the view to return the credit balance of the Reconciled Balance
instead of the value balance. This caused an incorrect value to be
displayed and used for client-side validation.

* Add documentation to methods in the accounts.forms module.
* Add non-field error output to the account_reconcile.html template.

* Refactor and document the javascript in the account_reconcile.html
  template.

Fixes #284: Account Reconcile Form Miscalculates Out of Balance Amount.