[#212] Add Account Autocompletes to Formsets
This modifies formsets to use Selectize.js widgets instead of the
default HTML Selects. To prevent excessive HTML generation, the options
for each Account field are stored in a single javascript variable.
Accounts are now shown if their description contains what the user has
typed(previously only the name was used for filtering). When the user
types in some text, an AJAX request is sent in case new Accounts have
been created.

* Add the djangoajax dependency.
* Add the all_accounts context processor & a global accounts_array
  javascript variable containing the options for the account selects.
* Add the accounts_query AJAX view to create an array of JSON objects
  for the account select options.