[#1000] Add Setting for Running on WAN
Add the REQUIRE_LOGIN setting to allow running the application on a WAN.
Enabling this setting modifies all pages(except the Login Page) to
require authentication, redirecting to the Login Page if the User is not
authenticated.

Pages can be made exempt from this check by adding them(or a regex that
includes them) to the REQUIRE_LOGIN_EXEMPT_URLS setting.

By default, the application sets REQUIRE_LOGIN to False. WAN sites
should set this to True in their local settings file.

Closes #1000